+++ Commands/Owners/remove-premium.js
@@ -19,14 +19,8 @@ module.exports = {
         User: member.id,
       },
       async (err, data) => {
-        if (!data)
-          return msg.channel
-            .send(`\`${member.user.username}\`Não tinha premium!`)
-            .then(msg.delete());
-        data.delete();
-        msg.channel
-          .send(`\`${member.user.username}\` já não tem premium`)
-          .then(msg.delete());
+        if (!data) return msg.channel.send(`\`${member.user.username}\`Não tinha premium!`).then(msg.delete());
+        data.delete();msg.channel.send(`\`${member.user.username}\` já não tem premium`).then(msg.delete());
       }
     );
   },
